# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: 15th MEU API
  # Unknown query parameters are ignored on every operation rather than
  # rejected: a request with an undeclared parameter is accepted, not a 400.
  # Tooling that flags this is seeing intended behavior, not a defect.
  description: >-
    Read-only access to 15th MEU roster and forum data. All endpoints require a
    bearer API key.


    ## Get your API key


    Obtain an API key from the 15th MEU API manager (members only). Enter it in the
    reference's authentication controls to try any endpoint. The docs send it as
    `Authorization: Bearer <key>`, so enter only the raw API key.
  version: dev
...
servers:
  - url: https://api.15thmeu.org
...
security:
  - bearer: []
paths:
  /api/v1/milpacs/profile/id/{userId}:
    get:
      operationId: MilpacService_GetProfile
      summary: Get given user milpac Profile
      description: >-
        Get milpac Profile data for a specific user. The path value is the
        milpac relation id, not the forum user id.
      tags: [milpacs]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
            description: 64-bit id as a decimal string.
        - name: username
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Error'
        "404":
          $ref: '#/components/responses/Error'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/profile/username/{username}:
    get:
      operationId: MilpacService_GetProfile2
      summary: Get given user milpac Profile
      description: Get milpac Profile data for a specific user by username.
      tags: [milpacs]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpac/discord/{discordId}:
    get:
      operationId: MilpacService_GetUserViaDiscordId
      summary: Get a Milpac profile via Discord ID
      description: Get all milpac Profile data for a given user via their Discord ID.
      tags: [milpacs]
      parameters:
        - name: discordId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpac/gamertag/{gamertag}:
    get:
      operationId: MilpacService_GetGamertagProfile
      summary: Get a Milpac profile via Gamertag
      description: Get all milpac Profile data for a given user via their console Gamertag.
      tags: [milpacs]
      parameters:
        - name: gamertag
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The full milpac profile.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/roster/{roster}:
    get:
      operationId: MilpacService_GetRoster
      summary: Get a given roster
      description: Get all milpac Profile data for a given roster.
      tags: [milpacs]
      parameters:
        - $ref: '#/components/parameters/RosterPath'
      responses:
        "200":
          description: Profiles keyed by milpac relation id.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/roster/{roster}/lite:
    get:
      operationId: MilpacService_GetLiteRoster
      summary: Get a given roster with minimal profile data
      description: >-
        Get basic milpac Profile data for a given roster, excluding records
        and awards.
      tags: [milpacs]
      parameters:
        - $ref: '#/components/parameters/RosterPath'
      responses:
        "200":
          description: Lite profiles keyed by milpac relation id.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiteRoster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/s1/uniforms/{roster}:
    get:
      operationId: MilpacService_GetS1UniformsRoster
      summary: Get a given roster tailored for S1 Uniforms
      description: Get the specified roster with data tailored for the S1 uniforms tool.
      tags: [milpacs]
      parameters:
        - $ref: '#/components/parameters/RosterPath'
      responses:
        "200":
          description: S1 uniforms profiles keyed by milpac relation id.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/S1UniformsRoster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/position/search/{positionQuery}:
    get:
      operationId: MilpacService_SearchByPosition
      summary: Search for profiles by position title
      # An empty profiles map is a valid result: {} means no current position
      # title contains the query.
      # The live route accepts multi-segment queries, so slashes inside the
      # query and a bare trailing slash also reach the handler. That is a form
      # an OpenAPI path template cannot express; this document records the
      # canonical single-segment form.
      description: >-
        Search for profiles by position title substring (SQL LIKE '%query%',
        case-insensitive, unanchored, against primary and secondary-capable
        position titles); returns lite profiles. An empty profiles map means
        no current position title contains the query. Terms follow the unit's
        own billet vocabulary (line leadership is "Section Leader" / "Platoon
        Leader"; there is no "Squad Leader" billet, so that query returns an
        empty map).
      tags: [milpacs]
      parameters:
        - name: positionQuery
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: >-
            Matching lite profiles keyed by milpac relation id; {} when
            nothing matches (200, never 404).
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiteRoster'
        "400":
          $ref: '#/components/responses/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/milpacs/ranks:
    get:
      operationId: MilpacService_GetAllRanks
      summary: Get all ranks
      description: Get a list of all ranks and their details.
      tags: [milpacs]
      responses:
        "200":
          description: Rank reference list.
          headers:
            Vary:
              description: >-
                Cache-keying hint: the server selects the body encoding from
                Accept-Encoding, so a shared cache must store the gzip and
                identity variants under separate keys.
              schema:
                type: string
                const: Accept-Encoding
            Cache-Control:
              description: >-
                Freshness signal: data may be up to 10 minutes stale.
                Consumers may treat a response as fresh for 10 minutes
                between polls.
              schema:
                type: string
                const: max-age=600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RanksResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    # http/bearer makes the docs UI auto-prepend "Bearer " so users paste only
    # the raw cav7_… key.
    bearer:
      type: http
      scheme: bearer
      description: >-
        15th MEU API key (cav7_…), sent as `Authorization: Bearer <key>`.
        Members generate a key from the API key manager at
        https://15thmeu.org/account/api-key.
        Requests without the Bearer scheme receive a plain-text 401 naming
        the expected header form; unknown keys receive a generic plain-text
        401; valid keys lacking the operation's scope receive a JSON 403.
...
  responses:
...
    Error:
      description: >-
        JSON error body. `code` is a numeric status code (3 InvalidArgument →
        400, 5 NotFound → 404, 7 PermissionDenied → 403, 13 Internal → 500).
        `message` is a human-readable description of the failure.
      headers:
        Vary:
          description: >-
            Cache-keying hint: the server selects the body encoding from
            Accept-Encoding, so a shared cache must store the gzip and
            identity variants under separate keys.
          schema:
            type: string
            const: Accept-Encoding
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
  schemas:
...
    Roster:
      type: object
      properties:
        profiles:
          type: object
          description: Profiles keyed by milpac relation id (64-bit ids as decimal strings).
          additionalProperties:
            $ref: '#/components/schemas/Profile'
      required: [profiles]
      additionalProperties: false
...